ci: rename graph-revive to revive-graphs and restructure command
- Rename workflow from graph-revive.yml to cron-revive-graphs.yml with clearer purpose
- Move CLI command from cli/commands/graph-revive/ to cli/commands/revive-graphs.ts
- Update command name from 'graph-revive' to 'revive-graphs' for consistency
- Fix import paths after file reorganization
- Simplify GitHub Action workflow with v1 setup action
- Update justfile recipe to match new command name
